+++ npm_dist.sh
@@ -1,25 +1,29 @@
 #!/bin/bash
+echo -e "[0/6]. Starting new lib-build..."
+echo -e "       Make sure:"
+echo -e "       - Version is bumped in package.json"
+echo -e "       - JS-banner is updated rollup.jsbanner.mjs"
+echo -e "       - CSS-banner is updated rollup.cssbanner.mjs \n"
+
 rm -rf ./dist
-echo "(1). Old dist removed"
+echo -e "[1/6]. Old dist removed \n"
 
 mkdir ./dist ./dist/dist ./dist/src ./dist/src/oltb
-echo "(2). New dist created"
+echo -e "[2/6]. New dist created \n"
 
 cp -r ./src/oltb/js ./dist/src/oltb
 cp -r ./src/oltb/scss ./dist/src/oltb
-echo "(3). Copied src files"
+echo -e "[3/6]. Copied src files \n"
 
 cp ./package.json ./dist/package.json
+cp ./LICENSE ./dist/LICENSE
 cp ./README.md ./dist/README.md
 cp ./README_INTERNAL.md ./dist/README_INTERNAL.md
-echo "(4). Copied documentation files"
+echo -e "[4/6]. Copied documentation files \n"
 
-echo "----------------------------------------------------------------"
-echo "(5). Make sure:"
-echo "     - Version is bumbed in package.json"
-echo "     - Version is bumbed in rollup.cssbanner.mjs"
-echo "     - Version is bumbed in rollup.jsbanner.mjs"
-echo "----------------------------------------------------------------"
-echo "(6). Next step:"
-echo "     - Create portable build '$ npm run build:lib'"
-echo "     - Publish new version to npmjs '$ bash npm_publish.sh'"
+npm run build:lib
+echo -e "[5/6]. Rollup created IIFE lib \n"
+
+echo -e "[6/6]. Next steps:"
+echo -e "     - Create and verify new example '$ npm link'"
+echo -e "     - Publish new version to npmjs '$ bash npm_publish.sh' \n"